+++ firmware-v2/src/animations/animations.h
@@ -34,6 +34,7 @@
 #include "squarral.h"
 #include "text.h"
 #include "tranquility.h"
+#include "gyrophare_fr.h"
 #include "warm_fade.h"
 #include "whirlwind.h"
 #include "zone.h"